modules/platforms/python/requirements.txt
pytest
Sphinx
attrib
